+++ packages/blockchain-wallet-v4-frontend/src/scenes/Earn/Table/MobileRow/MobileRow.styles.ts
