+++ packages/blockchain-wallet-v4-frontend/src/scenes/Earn/Table/SortableTable/SortableTable.styles.ts
@@ -1,0 +1,14 @@
+import { PaletteColors, TableContainer as Table } from '@blockchain-com/constellation'
+import styled from 'styled-components'
+
+export const TableContainer = styled(Table)`
+  width: 100%;
+  border-color: ${PaletteColors['grey-000']} !important;
+
+  & > thead > tr {
+    background-color: ${PaletteColors['grey-000']} !important;
+  }
+`
+export const ButtonContainer = styled.div`
+  min-width: 76px;
+`
